+++ src/main/java/com/example/acg_labs/parser/MtlParser.java
@@ -53,7 +53,7 @@ public class MtlParser {
                     case "Kd" -> mtlInfo.setKd(getDefaultColor(tokens));
                     case "Ks" -> mtlInfo.setKs(getDefaultColor(tokens));
 
-                    case "Ns" -> mtlInfo.setNs(Float.parseFloat(tokens[1]));
+                    case "Ns" -> mtlInfo.setNs(Integer.parseInt(tokens[1]));
 
                     case "map_Kd" -> mtlInfo.setMapKd(tokens[1]);
                     case "map_Ka" -> mtlInfo.setMapKa(tokens[1]);
